+++ CHANGELOG.md
@@ -4,7 +4,11 @@
 - Fixed incorrect default ore types in conversion config
 - Added option to disable logging startup rules, deny lists, etc.
 - Fixed swapping upgrade with stack count > 1 voiding extra items
+- Fixed admin key not locking if you don't own the drawer
 - Improved consistency with claimed/protected chunks (e.g. via FTB Chunks)
+- Added FTB variant of Personal Key to support team access
+- NEOFROGE/FABRIC: Added FTB Chunks support
+- NEOFORGE/FABRIC: Added FTB Teams support
 
 [13.9.3]
 - Fixed invalid gold_keyrings tag entries